fix : 가게평점 소숫점 누락 (int -> float 으로 조회되도록 변경)

+++ Spring/server/src/main/java/spring/reborn/domain/jjim/model/JjimStoreRes.java
@@ -15,5 +15,5 @@ public class JjimStoreRes {
     private String storeName;
     private String storeImage;
     private String storeCategory;
-    private int storeScore;
+    private float storeScore;
 }
